requirements.txt
requests
psycopg2-binary
pandas
python-dotenv
dvc 
dvc-s3
scikit-learn 
scipy 
numpy
mlflow==2.16.2
xgboost
fastapi
uvicorn
pydantic
streamlit
apache-airflow